rust: clean up FPIOA interface
Remove need for .into() everywhere by accepting everything
that goes into usize in the appropriate methods.

+++ rust/uart-passthrough/src/main.rs
@@ -29,10 +29,10 @@ fn main() -> ! {
     // Configure UART1 (→WIFI)
     sysctl::clock_enable(sysctl::clock::UART1);
     sysctl::reset(sysctl::reset::UART1);
-    fpioa::set_function(io::WIFI_RX as u8, fpioa::function::UART1_TX);
-    fpioa::set_function(io::WIFI_TX as u8, fpioa::function::UART1_RX);
-    fpioa::set_function(io::WIFI_EN as u8, fpioa::function::GPIOHS8);
-    fpioa::set_io_pull(io::WIFI_EN as u8, fpioa::pull::DOWN);
+    fpioa::set_function(io::WIFI_RX, fpioa::function::UART1_TX);
+    fpioa::set_function(io::WIFI_TX, fpioa::function::UART1_RX);
+    fpioa::set_function(io::WIFI_EN, fpioa::function::GPIOHS8);
+    fpioa::set_io_pull(io::WIFI_EN, fpioa::pull::DOWN);
     gpiohs::set_pin(8, true);
     gpiohs::set_direction(8, gpio::direction::OUTPUT);
     let mut wifi_serial = p.UART1.constrain(DEFAULT_BAUD.bps(), &clocks);
